refactor: 권한을 확인하는 확장함수 Permission 클래스로 이동
