--- app/src/main/java/woowacourse/movie/service/ContextExtension.kt
@@ -1,18 +1,0 @@
-package woowacourse.movie.service
-
-import android.Manifest
-import android.content.Context
-import android.content.pm.PackageManager
-import android.os.Build
-import androidx.core.content.ContextCompat.checkSelfPermission
-
-object ContextExtension {
-
-    fun Context.checkNotificationSelfPermission(): Boolean {
-        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            checkSelfPermission(
-                this, Manifest.permission.POST_NOTIFICATIONS
-            ) == PackageManager.PERMISSION_GRANTED
-        } else false
-    }
-}
