+++ app/src/main/java/woowacourse/movie/service/PermissionManager.kt
@@ -1,9 +1,12 @@
 package woowacourse.movie.service
 
 import android.Manifest
+import android.content.Context
+import android.content.pm.PackageManager
 import android.os.Build
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
+import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 
 object PermissionManager {
@@ -30,5 +33,13 @@ object PermissionManager {
                 onGranted()
             }
         }
+    }
+
+    fun Context.checkNotificationSelfPermission(): Boolean {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            ContextCompat.checkSelfPermission(
+                this, Manifest.permission.POST_NOTIFICATIONS
+            ) == PackageManager.PERMISSION_GRANTED
+        } else false
     }
 }
